requirements.txt
 python-binance